Initial commit to remove some boilerplate.
Want all the models in the models/ folder to be clean of repetitive code.
We created a base abstract that takes care of defining the id, toMap, insert,
update, and delete methods. We also added a string getter for tableName because
we needed to define that variable for all the abstract methods. Finally, we set
id on the await insert line to update the object with the id during insert.