Change all datetime strings to actual DateTime object
Changed all Strings to DateTime. Then when creating objects we used DateTime.parse(). There was some
goofy stuff we needed to do due to the fact that sometimes the strings were null. More specifically
for times that could actually be null, they were actually a string 'null'. For other cases, we
wanted the time to be optional but then initialize to now. To get that done, we had to move the time
initialization steps to toMap().

Note that A??=B is the same as A = A ?? B. A is A unless it's null, then it's B.