Create fromSelf and fromOther commands for every model to simple list generation.
Our code was starting to look really ugly from a ton of nested instantiation commands occuring in
list generation. I had an idea to help clean this up by functionalizing those commands. The next step
from here would be to look into making the SELECT statements in the raw queries more standardized. If
that is possible, we could move the entire list generation step to the base class.