Move to using a migrations-based system for upping/downing the database via the sqlx cli library.
Using migrations seems to be typical best practices from my cursory searchings on the internet. Expecially with the
sqlx library. So, it would be best to try to learn and use it. Plus, it helps us make testing easier. The sqlx
testing library will automatically create and destory the database for you for each integration test so it starts
fresh each time. This helps us not need to write a bunch of boiler plate to clean the database after each test.