include service to fetch all score types
we have a basic test to pull in a score type for "softball".
currently there is a bug where sport names in snake case are not
deserialized properly into the uppercase enum option, so we had
to include the '#[allow(non_camel_case_types)]' macro to permit
the lowercase options to support this. it seems weird because we
were able to work with Gender and TrophyType in lowercase before,
but what's different now is that the enum is being deserialized
through the sqlx::Path extractor and not sqlx::Json. Maybe a bug?