Work in progress: implement game services and tests
Currently running into an issue where rust/postgres wont serialize/deserialize OffsetDateTime or timestamptz correctly. Not sure
if it is an issue in how we defined things or if we need to add ugly parsing and formatting code. It feels like it should be smart
enough to handle that. We also cant figure out a way to do debugging due to rust-analyzer not working in vscode at the moment