fix: deserialize iso8601 datetimes in request body
this wasn't working because, for some reason we still dont' understand,
you need to use the T separator in the datetime format instead of the
traditional " ". very stupid, i know...

also, when using a payload as bytes, you need to use `set_payload` and
not `set_json`. we also don't need the `serde_with` macros on `Game`
because it seems to do this automatically. we just need it on the field
in `CreateGameBody`.