Add services and testing for Score.
Followed the template we've got going for previous tables in the database. No real issues
getting this one up and running. We might want to look into getting all the fixtures to
call the fixtures they depond on automatically, instead of setting it in each test because
things are looking pretty repetitive the farther we go.