Add services for Attendance.
Need to create tests next. We decide to link teammate instead of player. Thinking on it,
it would probably be best to use teammate for score as well. We can save that for another
commit if it starts to make sense to do. There are also a lot more endpoints and services
we could be creating but that might not be useful to the frontend. So, we are saving
them for when we start to work onthe frontend and realize it would be nice to have a new
service.