Delete every thing we did with flutter...
We had only done backend stuff with flutter. And we already moved to using rust instead.
In the meantime, we also decided that maybe flutter isn't such a great idea. Instead, we
will move forward with keeping as native as possible (using Kotlin MPM since Kotlin is
native to android, and then we will just use Swift for iOS specifics that MPM wont let
us do).