Switch to material3 since it's new and shiny.
While googling around I noticed that material3 has been released and is stable.
I tried changing this template repo to use it to see if it would still work and it did.
I also test on android and got ios working too. For ios, you need to clean the pod cache
and then pod reinstall for the m-series architecture. Then cd into the iosApp folder to run
the test configuration. No need to open up xcode.