throw together some basic functionality in the app
Bottom bar toggles what is shown in main view and the top bar.
The roster view has a basic list of cards with an image and text.
Floating button is there too.
Weird that we have to mark functions with all these experimental things?
Ideally there won't be so many. Right now we have to use it on every function.