Hacky way to get iOS to build by moving all UI logic to android.
I'm pretty sure we are going about this in a goofy way. All UI logic should probably be platform specific since a lot of
the fancy material3 components do NOT work on iOS and we will need to code swift. It would probably make sense to move
forward with only keeping business logic in common.