connect frontend to backend
this doesn't work yet because of a networking issue. essentially, we
need to forward traffic from the local machine to the emulator so that
the running application can communicate with the backend running on the
computer.

we had to include some tags in the AndroidManifest to allow the
application to communicate over HTTP instead of HTTPS.

we also use this library called ktor which we will continue to use as an
http client to query our backend.