write python script to generate dummy data
currently does not generate a sql file to load into the database, but it
has all of the logic that relates our models. eventually we want to
pivot to writing this script in rust somehow using the models we have
already defined (or maybe even in raw sql?).

also provided a requirements.txt in order to initialize the necessary
virtual environment to run this script.